trials/sutra-ai old/backend-python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.6
faiss-cpu==1.7.4
pypdf==3.17.4
beautifulsoup4==4.12.3
requests==2.31.0
numpy==1.26.3
pydantic==2.5.3
python-multipart==0.0.6
pillow==10.2.0